Move the output name overlay to the bottom left
The output name overlay renders on the top left by default. At the same
time, wdisplays renders the output previews on its top left corner too.

The results is that the overlay covered the output previews, making
seeing it and interacting with it tricky.

Moving the output name to the bottom left makes sure there's never any
overlap between it and the draggable previews shows for each output.